requirements.txt
altgraph==0.17.4
anyio==4.6.2.post1
auto-py-to-exe==2.46.0
beautifulsoup4==4.13.3
bottle==0.13.2
bottle-websocket==0.2.9
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clr_loader==0.2.7.post0
dnspython==2.7.0
Eel==0.18.1
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
motor==3.6.0
oauthlib==3.2.2
packaging==24.2
pefile==2023.2.7
ply==3.11
proxy_tools==0.1.0
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pymongo==4.9.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6-Qt6==6.9.0
PyQt6-WebEngine-Qt6==6.9.0
PyQt6_sip==13.10.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
python-telegram-bot==21.6
pythonnet==3.0.5
pytube==15.0.0
pywebview==5.4
pywin32-ctypes==0.2.3
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
setuptools==78.1.0
sniffio==1.3.1
soupsieve==2.6
Telethon==1.37.0
tweepy==4.14.0
typing_extensions==4.13.0
urllib3==2.2.3
websockets==15.0.1
zope.event==5.0
zope.interface==7.2
PyQt5==5.15.10
PyQtWebEngine==5.15.10
pdoc==15.0.3